# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Engepe\u00e7as - Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/purchaseHistIncomingInvoices": {
            "get": {
                "tags": [
                    "PurchaseHistIncomingInvoice"
                ],
                "summary": "Get a listing of the PurchaseHistIncomingInvoices.",
                "description": "Get all PurchaseHistIncomingInvoices",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "PHO_Id",
                        "in": "query",
                        "description": "Ordem de Compra",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade limite de exibi\u00e7\u00e3o",
                        "required": false,
                        "type": "integer",
                        "default": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "P\u00e1gina a ser exibida",
                        "required": false,
                        "type": "integer",
                        "default": "1"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "HRD_Quantidade"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Dire\u00e7\u00e3o da ordena\u00e7\u00e3o do retorno",
                        "required": false,
                        "type": "string",
                        "default": "DESC"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "Informe a sele\u00e7\u00e3o de campos que devem retornar da consulta separados por virgula",
                        "required": false,
                        "type": "string",
                        "default": "id, PHO_Id, HRD_T014_Id, HRD_Quantidade,HRD_Valor_Custo_Unitario,HRD_Flag_Cancelado,HRD_Data_Lancamento"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pesquise por qualquer campo, ao usar este campo as outras consultas ser\u00e3o desconsideradas",
                        "required": false,
                        "type": "string",
                        "default": "Busca"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/PurchaseHistIncomingInvoice"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "PurchaseHistIncomingInvoice"
                ],
                "summary": "Store a newly created PurchaseHistIncomingInvoice in storage",
                "description": "Store PurchaseHistIncomingInvoice",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "PHO_Id",
                        "in": "query",
                        "description": "Ordem de Compra",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Item de Nota Fiscal que deve ser armazenado.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/PurchaseHistIncomingInvoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PurchaseHistIncomingInvoice"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            }
        },
...
        "/purchaseHistIncomingInvoices/{id}/purchaseHistOrders": {
            "get": {
                "tags": [
                    "PurchaseHistIncomingInvoice"
                ],
                "summary": "Listar Itens com a Ordem de Compra Especifica. ",
                "description": "Entre com o Registro.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Codigo da Nota de Entrada",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PurchaseHistIncomingInvoice"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            }
        },
...
        "/purchaseHistIncomingInvoices/{HRD_T014_Id}/getpurchaseHistIncomingInvoices": {
            "get": {
                "tags": [
                    "PurchaseHistIncomingInvoice"
                ],
                "summary": "Exibir a Item de Nota Fiscal especificado Pelo T014_Id (Numero do Item na Nota Fiscal)",
                "description": "Seleciona Item de Nota Fiscal",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "HRD_T014_Id",
                        "in": "path",
                        "description": "Codigo do Item na Nota Fiscal",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opera\u00e7\u00e3o realizada com sucesso.",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PurchaseHistIncomingInvoice"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "EngepecasAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "PurchaseHistIncomingInvoice": {
            "required": [
                "PHO_Id",
                "HRD_T014_Id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "PHO_Id": {
                    "description": "PHO_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "HRD_T014_Id": {
                    "description": "HRD_T014_Id",
                    "type": "integer",
                    "format": "int32"
                },
                "HRD_Quantidade": {
                    "description": "HRD_Quantidade",
                    "type": "integer",
                    "format": "int32"
                },
                "HRD_Valor_Custo_Unitario": {
                    "description": "HRD_Valor_Custo_Unitario",
                    "type": "number",
                    "format": "number"
                },
                "HRD_Flag_Cancelado": {
                    "description": "S => SIM CANCELADA | N => NAO CANCELADA",
                    "type": "string"
                },
                "HRD_Data_Lancamento": {
                    "description": "Data de entrada nf compra",
                    "type": "string",
                    "format": "date-time"
                },
                "created_by": {
                    "description": "created_by",
                    "type": "integer",
                    "format": "int32"
                },
                "updated_by": {
                    "description": "updated_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_by": {
                    "description": "deleted_by",
                    "type": "integer",
                    "format": "int32"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "EngepecasAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}